requirements.txt
requests
python-dotenv
dateparser
ollama